+++ crates/tracel-cloud/src/backend/station.rs
@@ -1,0 +1,12 @@
+use burn_central_client::StationClient;
+
+#[derive(Debug, Clone)]
+pub struct StationBackend {
+    pub client: StationClient,
+}
+
+impl StationBackend {
+    fn new(client: StationClient) -> Self {
+        Self { client }
+    }
+}
